Test that xterm-256color ends up the same
This is likely to *fail* over time. The xterm-256color entry does
change, but not really in ways we care about. (even if it does change
for sequences we use it's often in some annoying "oh and let clear
erase scrollback too" way)

I'll probably end up removing this, but I would like to see if it
matches *now*.